I refactored the movemen.py module into a package, and began refactoring the Movement class.
