+++ pterasoftware/movements/__init__.py
@@ -1,0 +1,31 @@
+"""This package contains the movement classes and their helper functions.
+
+This package contains the following subpackages:
+    None
+
+This package contains the following directories:
+    None
+
+This package contains the following modules:
+    __init__.py: This module is this package's initialization script.
+
+    airplane_movement.py: This module contains the AirplaneMovement class.
+
+    functions.py: This module contains useful functions for the movement classes.
+
+    movement.py: This module contains the Movement class.
+
+    operating_point_movement.py: This module contains the OperatingPointMovement class.
+
+    wing_cross_section_movement.py: This module contains the WingCrossSectionMovement
+    class.
+
+    wing_movement.py: This module contains the WingMovement class.
+"""
+
+import pterasoftware.movements.airplane_movement
+import pterasoftware.movements.functions
+import pterasoftware.movements.movement
+import pterasoftware.movements.operating_point_movement
+import pterasoftware.movements.wing_cross_section_movement
+import pterasoftware.movements.wing_movement
